+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.15.2
  * Group the padding digits to the left of the number, if present. e.g. 00,001.
+ * Update the SDK constraint for Dart 2.0 dev versions.
 
 ## 0.15.1
  * Use the platform.locale API to get the OS platform.